MuxError::BufferFull + Muxer accessors: usize -> u64 (FFI portability)
Audit theme H9. Counters in published shapes are u64 for predictable
width on 32-bit ARM (UniFFI/JNI/PyO3 binding targets). Internal usize
values cast at the public boundary.

  - MuxError::BufferFull::capacity_packets:   usize -> u64
  - Muxer::pending_packets():                 usize -> u64
  - Muxer::capacity_packets():                usize -> u64

Construction sites (4 in mpegts::mux) bridge with `as u64` from the
internal `MuxerConfig::buffer_packets: usize` (a Rust-API knob that
stays usize per Phase 3.5.4 inventory). Caller in MuxSender's
back-pressure tracer (`maybe_warn_backpressure`) is unchanged — both
operands flip from usize to u64 together so arithmetic stays clean.

+++ crates/tst-core/src/error.rs
@@ -184,7 +184,7 @@ pub enum MuxError {
     InvalidNal,
 
     #[error("muxer packet buffer is full ({capacity_packets} packets); drain via pull and retry")]
-    BufferFull { capacity_packets: usize },
+    BufferFull { capacity_packets: u64 },
 
     /// KLV blob exceeds the 16-bit `PES_packet_length` ceiling.
     ///
